Fixed Bug in TimedWrites
The loop over TimedWrites incorrectly erased iterators inside a for-loop.  For some reason this never caused an issue on the ESP32, but crashed on the ESP32-C3.  Solution is to change the for-loop to a while-loop with proper handling of the iterator when an element is deleted.  This appears to fix the problem.