Updated "Other Examples" and added SPAN_ACCESSORY() Macro
Need to revisit Television sketch - the latest iOS seemed to have disabled the ability to dynamically set visibility of input sources.  Checking/Un-Checking the visibility radio buttons either during pairing, or after pairing on the settings screen, seems to be ignored by the Home App (the same input sources are shown in the selector regardless of any changes made).  However, dynamically changing the name of an input source seems to work fine.