Added optional AccessoryFlags Characteristic to AccessoryInformation Service
This optional flag does not seem to have any affect on the Home App and does not produce any messages indicating that the device needs additional setup.  Was hopeful this would have prevented the "Not Supported" message on a Home App Tile that is configured as a bridge, but without any other Accessories.