Move all SpanPoint code into separate HomePoint.cpp and HomePoint.h files
This enables SpanPoint to operate as a standalone library with no reliance on anything in HomeSpan.

To Do: migrate all HomePeer code into SpanPoint.  This will yield a universal standalone HomePoint library that can be used for BOTH the receiver and the sender (which automatically means two-way communication is always possible)