Moved "Other Examples" into "examples" folder
Noticed that the Arduino IDE 2.0 interface was not displaying Other Examples.

This is because Other Examples was not under the "examples" folder.  Not sure why this was the case - I should have originally put it there (?)  Regardless, once moved into the examples folder, Other Examples now properly shows up in both Arduino IDE version 1 and version 2.