Added homeSpan.refreshStatusDevice()
Provides for additional control over Status LED when a generic device has been used.  If the color of the device is changed, the change will only take effect when the LED is next turned on.  If the LED is blinking, the color changes upon the next blink.  But if the LED is ON, there is no next blink and color will not change.

This method check the status of the LED, and if it is ON, is calls the on() function again, which will result in an update of the color.  The method does nothing if the LED is OFF or BLINKING.