Updated mapping Characters to allow for "01234" and "-"
Use "01234" as alternatives to "RGBCW".  Use "-" as a blank placeholder for unconnected Pixel wire, as typically used in Warm/Cool-White LEDs, such as "WC-" or "-CW".  Note "-" simply maps to R (same as 0) but does not matter since wire is not connected - we just need to ensure a byte is transmitted.