Updated TOUCH code in PushButton for 3.1.0 compatability
In 3.1.0 the macros SOC_TOUCH_VERSION_1 and SOC_TOUCH_VERSION_2 were deleted and replaced with a single SOC_TOUCH_SENSOR_VERSION macro that can be set to 1, 2, or 3 (for the P4 chip - not supported by HomeSpan)

This NOT-BACKWARDS-COMPATIBLE change breaks the TOUCH logic in Pushbutton.  Fix is to recognize BOTH forms of the macro ensure both forward AND backwards compatibility.