Added setCompileTime() method and HS_SET_COMPILE_TIME macro
Allows user to set the compile time of the sketch itself (as opposed to the last compile time of the HomeSpan library, as previously shown during HomeSpan boot-up).  If unspecified, compile time will be displayed as "N/A" in Serial Monitor and not included at all in top WebLog table.

Macro HS_SET_COMPILE_TIME is user-friendly way of setting compile time based on predefined __DATE__ and __TIME__ macros.