update watchdog logic to include yieldIfNecessary()
Linked to yieldIfNecessary() in main.cpp of Ardiuno-ESP32 if single-core.  If mult-core, yieldIfNecessary() is defined to be identical to that of Arduino-ESP32.

To do: create default idle-core-mask and use in enableWatchdog() so it adds to ARDUINO_CORE, as well as when autopolling starts.  This allows watchdog to protect against problems in set-up before autoPolling even starts.