Do NOT remove backslashes from STRING JSON
Continue to remove backslashes from DATA and TLV_ENC JSON, but not from STRING JSON to allow use of escaped characters to be properly received in JSON STRING-based Characteristics (such as a quote).

This means an escaped forward-slash will retain the backslash when used in a STRING Characteristic, but the backslash will be removed from DATA and TLV8 Characteristics.  HomeKit will accept an unescaped forward-slash (as required by JSON spec) but will add optional escape when transmitting, which is fine for STRINGS, but corrupts DATA and TLV8 base-64 encoding.