Added ability to directly enter SHA256 or MD5 hash of OTA password
Prefix password with 0x to have it interpreted as a hash instead of plain-text.
Works with enableOTA() from within a sketch as well as when setting the OTA password with the "O" CLI command.

HomeSpan validates hash to make sure it contains only exactly 32 (for MD5) or exactly 64 (for SHA256) hexidecimal characters.  If not, HomeSpan interprets as plain-text if total size (including initial 0x) is 32 characters or less.  If more than 32 characters in total, a warning is thrown and the OTA password is NOT updated.

Any hash entered will be automatically converted to all lower-case when used/stored.